# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BBS论坛用户管理API
  description: |
    BBS论坛系统的用户认证与管理接口文档
    
    ## 功能概述
    本API提供用户注册、登录、个人信息管理等核心功能，支持多种认证方式。
    
    ## 认证机制
    - JWT Token认证：登录成功后返回JWT Token，需在请求头中携带 `Authorization: Bearer <token>`
    - 三种登录方式：
      1. 用户名密码登录：传统的用户名+密码认证
      2. 邮箱验证码登录：通过邮箱接收验证码进行认证
      3. 微信小程序登录：使用微信小程序授权码登录
    - 用户注册支持：邮箱验证码注册、微信小程序注册
    
    ## 错误处理
    所有接口均返回统一格式的响应，包含状态码、消息和数据。错误时会返回对应的错误码和错误信息。
  version: 0.1.0
  contact:
    name: BBS开发团队-后端组
    email: leotanstudio@gmail.com
servers:
  - url: http://bbs.leocoding.onliine
    description: 测试环境
paths:
...
  /user/{username}:
    get:
      tags:
        - 用户管理
      summary: 根据用户名获取用户信息
      description: |
        获取指定用户名的用户公开信息。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            example: "johndoe"
          description: 用户名
      responses:
        '200':
          description: 获取用户信息成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetails'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /posts/draft:
    get:
      tags:
        - 帖子管理
      summary: 获取当前用户帖子草稿
      description: |
        获取当前登录用户最新保存的帖子草稿。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: 获取草稿成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /posts/{postId}:
    parameters:
      - name: postId
        in: path
        required: true
        schema:
          type: integer
        description: 帖子ID
    get:
      tags:
        - 帖子管理
      summary: 获取帖子详情
      description: |
        根据帖子ID查看已发布帖子的完整信息。
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: 帖子不存在或未发布
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 帖子管理
      summary: 删除帖子
      description: |
        删除指定帖子。
        
        **权限要求：** 需要登录且必须是帖子作者或具有管理权限
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: 删除结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限删除该帖子
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /posts/user/{userId}:
    get:
      tags:
        - 帖子管理
      summary: 查看用户发布的帖子
      description: |
        根据用户ID分页查询该用户发布的公开帖子。
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从0开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: 每页数量
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 排序字段，格式示例：`createdAt,desc`
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPageResponse'
  /posts/{postId}/like:
    post:
      tags:
        - 帖子互动
      summary: 点赞帖子
      description: |
        为指定帖子点赞，返回点赞后的状态。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
      responses:
        '200':
          description: 点赞成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 帖子不存在或不可操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 帖子互动
      summary: 取消点赞帖子
      description: |
        取消当前用户对指定帖子的点赞。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
      responses:
        '200':
          description: 取消点赞成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 帖子不存在或不可操作
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/{postId}/like/status:
    get:
      tags:
        - 帖子互动
      summary: 查询帖子点赞状态
      description: |
        判断当前登录用户是否已点赞指定帖子。未登录时返回`false`。
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /posts/{postId}/like/count:
    get:
      tags:
        - 帖子互动
      summary: 获取帖子点赞数
      description: |
        返回指定帖子的总点赞数量。
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegerResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /comments:
    post:
      tags:
        - 评论管理
      summary: 发表评论
      description: |
        在帖子下创建新的评论或回复。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '200':
          description: 评论创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCommentEntityResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限评论
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 帖子或父评论不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/direct:
    post:
      tags:
        - 评论管理
      summary: 直接发表评论
      description: |
        直接发布仅包含文字内容的评论或回复，跳过草稿箱流程。

        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '200':
          description: 评论发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCommentEntityResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 帖子或父评论不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/{commentId}:
    parameters:
      - name: commentId
        in: path
        required: true
        schema:
          type: integer
        description: 评论ID
    get:
      tags:
        - 评论管理
      summary: 获取评论详情
      description: |
        根据评论ID获取评论内容及互动信息。
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponseWrapper'
        '404':
          description: 评论不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - 评论管理
      summary: 编辑评论
      description: |
        更新评论内容，仅允许作者本人操作。
        
        **权限要求：** 需要登录并拥有该评论
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: 编辑成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCommentEntityResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限编辑该评论
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 评论不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 评论管理
      summary: 删除评论
      description: |
        删除指定评论，需作者本人或拥有管理权限。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限删除该评论
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 评论不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/post/{postId}:
    get:
      tags:
        - 评论管理
      summary: 分页获取帖子评论
      description: |
        获取帖子下的评论列表（包含互动信息），支持分页。
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从0开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: 每页数量
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 排序字段，格式示例：`createdAt,asc`
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPageResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/{commentId}/replies:
    get:
      tags:
        - 评论管理
      summary: 分页获取评论回复
      description: |
        获取某条评论的回复列表（包含互动信息），支持分页。
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
          description: 父评论ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从0开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: 每页数量
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 排序字段，格式示例：`createdAt,asc`
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPageResponse'
        '404':
          description: 评论不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/user/{userId}:
    get:
      tags:
        - 评论管理
      summary: 分页获取用户评论
      description: |
        获取指定用户发表的评论列表，仅本人可查看完整数据。
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: 用户ID
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: 页码，从0开始
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: 每页数量
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: 排序字段，格式示例：`createdAt,desc`
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentPageResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限查看他人评论列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/draft:
    get:
      tags:
        - 评论管理
      summary: 获取评论草稿
      description: |
        获取当前用户针对指定帖子的评论草稿。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      parameters:
        - name: postId
          in: query
          required: true
          schema:
            type: integer
          description: 草稿对应的帖子ID
        - name: parentCommentId
          in: query
          required: false
          schema:
            type: integer
          description: 草稿回复的父评论ID，顶级评论留空
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCommentEntityResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 草稿不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - 评论管理
      summary: 编辑评论草稿
      description: |
        更新评论草稿的目标与内容。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftCommentEditorRequest'
      responses:
        '200':
          description: 草稿更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCommentEntityResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限编辑草稿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 草稿不存在或目标无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - 评论管理
      summary: 清空评论草稿
      description: |
        删除当前用户的评论草稿。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: 清空成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 草稿不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/draft/publish:
    post:
      tags:
        - 评论管理
      summary: 发布评论草稿
      description: |
        发布当前用户的评论草稿。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCommentEntityResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限发布草稿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 草稿不存在或状态不允许发布
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/draft/exists:
    get:
      tags:
        - 评论管理
      summary: 检查评论草稿是否存在
      description: |
        返回当前用户是否存在未发布的评论草稿。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/{commentId}/like:
    post:
      tags:
        - 评论管理
      summary: 点赞或取消点赞评论
      description: |
        切换评论的点赞状态，返回当前状态。
        
        **权限要求：** 需要登录
      security:
        - bearerAuth: [ ]
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
          description: 评论ID
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 评论不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/{commentId}/like/status:
    get:
      tags:
        - 评论管理
      summary: 查询评论点赞状态
      description: |
        判断当前登录用户是否点赞指定评论。未登录时返回`false`。
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
          description: 评论ID
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '404':
          description: 评论不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/{commentId}/like/count:
    get:
      tags:
        - 评论管理
      summary: 获取评论点赞数
      description: |
        返回评论被点赞的总次数。
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
          description: 评论ID
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LongResponse'
        '404':
          description: 评论不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /comments/post/{postId}/count:
    get:
      tags:
        - 评论管理
      summary: 统计帖子评论数量
      description: |
        返回帖子下评论（含回复）的总数量。
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
          description: 帖子ID
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegerResponse'
        '404':
          description: 帖子不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Token认证，格式为：Bearer <token>
        
        获取方式：
        1. 用户名密码登录：POST /auth/login
        2. 邮箱验证码登录：POST /auth/login-with-code
        3. 微信小程序登录：POST /auth/login-with-wechat
        
        Token会在登录成功后通过响应头Authorization和响应体data.token同时返回。
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: 错误代码
          example: 10001
        message:
          type: string
          description: 错误消息
          example: "Invalid parameters"
        data:
          oneOf:
            - type: "null"
            - type: object
              additionalProperties:
                type: string
          description: 错误详细信息，验证失败时包含字段级错误
          example:
            username: "用户名长度必须在2-50个字符之间"
            password: "密码长度必须在8-20个字符之间"
...
    PostResponse:
      type: object
      properties:
        code:
          type: integer
          example: 10000
        message:
          type: string
          example: "Success"
        data:
          $ref: '#/components/schemas/PostDetail'
    PostPageResponse:
      type: object
      properties:
        code:
          type: integer
          example: 10000
        message:
          type: string
          example: "Success"
        data:
          $ref: '#/components/schemas/PostPage'
    BooleanResponse:
      type: object
      properties:
        code:
          type: integer
          example: 10000
        message:
          type: string
          example: "Success"
        data:
          type: boolean
          example: true
...